IS-1974: Move aktivitetskrav folder to /sider folder
