+++ src/sider/aktivitetskrav/AktivitetskravAlertstripe.tsx
