+++ src/sider/aktivitetskrav/AktivitetskravContainer.tsx
@@ -5,7 +5,7 @@ import Sidetopp from "@/components/Sidetopp";
 import { useAktivitetskravQuery } from "@/data/aktivitetskrav/aktivitetskravQueryHooks";
 import { useOppfolgingstilfellePersonQuery } from "@/data/oppfolgingstilfelle/person/oppfolgingstilfellePersonQueryHooks";
 import SideLaster from "@/components/SideLaster";
-import { AktivitetskravSide } from "@/components/aktivitetskrav/AktivitetskravSide";
+import { AktivitetskravSide } from "@/sider/aktivitetskrav/AktivitetskravSide";
 import { NotificationProvider } from "@/context/notification/NotificationContext";
 import UtdragFraSykefravaeret from "@/components/utdragFraSykefravaeret/UtdragFraSykefravaeret";
 import * as Tredelt from "@/sider/TredeltSide";
