+++ src/sider/aktivitetskrav/vurdering/UnntakAktivitetskravSkjema.tsx
@@ -2,42 +2,43 @@ import React from "react";
 import {
   AktivitetskravStatus,
   CreateAktivitetskravVurderingDTO,
-  OppfyltVurderingArsak,
+  UnntakVurderingArsak,
   VarselType,
 } from "@/data/aktivitetskrav/aktivitetskravTypes";
-import { oppfyltVurderingArsakTexts } from "@/data/aktivitetskrav/aktivitetskravTexts";
-import { SkjemaHeading } from "@/components/aktivitetskrav/vurdering/SkjemaHeading";
+import { unntakVurderingArsakTexts } from "@/data/aktivitetskrav/aktivitetskravTexts";
+import { SkjemaHeading } from "@/sider/aktivitetskrav/vurdering/SkjemaHeading";
 import { SkjemaInnsendingFeil } from "@/components/SkjemaInnsendingFeil";
 import { useVurderAktivitetskrav } from "@/data/aktivitetskrav/useVurderAktivitetskrav";
 import {
   AktivitetskravSkjemaValues,
   VurderAktivitetskravSkjemaProps,
-} from "@/components/aktivitetskrav/vurdering/vurderAktivitetskravSkjemaTypes";
-import { SkjemaFieldContainer } from "@/components/aktivitetskrav/vurdering/SkjemaFieldContainer";
+} from "@/sider/aktivitetskrav/vurdering/vurderAktivitetskravSkjemaTypes";
+import { SkjemaFieldContainer } from "@/sider/aktivitetskrav/vurdering/SkjemaFieldContainer";
 import { useForm } from "react-hook-form";
 import { Button, Radio, RadioGroup } from "@navikt/ds-react";
 import BegrunnelseTextarea, {
   begrunnelseMaxLength,
-} from "@/components/aktivitetskrav/vurdering/BegrunnelseTextarea";
-import { useAktivitetskravNotificationAlert } from "@/components/aktivitetskrav/useAktivitetskravNotificationAlert";
+} from "@/sider/aktivitetskrav/vurdering/BegrunnelseTextarea";
+import { useAktivitetskravNotificationAlert } from "@/sider/aktivitetskrav/useAktivitetskravNotificationAlert";
 import { useAktivitetskravVarselDocument } from "@/hooks/aktivitetskrav/useAktivitetskravVarselDocument";
 
 const texts = {
-  title: "Er i aktivitet",
+  title: "Sett unntak fra aktivitetskravet",
   arsakLegend: "Årsak (obligatorisk)",
-  begrunnelseLabel: "Begrunnelse",
+  begrunnelseLabel: "Begrunnelse (obligatorisk)",
   missingArsak: "Vennligst angi årsak",
+  missingBegrunnelse: "Vennligst angi begrunnelse",
   lagre: "Lagre",
 };
 
-export interface OppfyltAktivitetskravSkjemaValues
-  extends AktivitetskravSkjemaValues {
-  arsak: OppfyltVurderingArsak;
-}
-
 const defaultValues = { begrunnelse: "", arsak: undefined };
 
-export const OppfyltAktivitetskravSkjema = ({
+export interface UnntakAktivitetskravSkjemaValues
+  extends AktivitetskravSkjemaValues {
+  arsak: UnntakVurderingArsak;
+}
+
+export const UnntakAktivitetskravSkjema = ({
   aktivitetskravUuid,
 }: VurderAktivitetskravSkjemaProps) => {
   const {
@@ -46,24 +47,22 @@ export const OppfyltAktivitetskravSkjema = ({
     formState: { errors },
     handleSubmit,
     reset,
-  } = useForm<OppfyltAktivitetskravSkjemaValues>({
-    defaultValues,
-  });
+  } = useForm<UnntakAktivitetskravSkjemaValues>({ defaultValues });
   const vurderAktivitetskrav = useVurderAktivitetskrav(aktivitetskravUuid);
   const { getVurderingDocument } = useAktivitetskravVarselDocument();
   const { displayNotification } = useAktivitetskravNotificationAlert();
 
-  const submit = (values: OppfyltAktivitetskravSkjemaValues) => {
-    const { arsak, begrunnelse } = values;
-    const status = AktivitetskravStatus.OPPFYLT;
+  const submit = (values: UnntakAktivitetskravSkjemaValues) => {
+    const status = AktivitetskravStatus.UNNTAK;
+    const { begrunnelse, arsak } = values;
     const createAktivitetskravVurderingDTO: CreateAktivitetskravVurderingDTO = {
       status,
       arsaker: [arsak],
       beskrivelse: begrunnelse,
       document: getVurderingDocument({
-        varselType: VarselType.OPPFYLT,
-        arsak,
         begrunnelse,
+        arsak,
+        varselType: VarselType.UNNTAK,
       }),
     };
     vurderAktivitetskrav.mutate(createAktivitetskravVurderingDTO, {
@@ -84,7 +83,7 @@ export const OppfyltAktivitetskravSkjema = ({
           legend={texts.arsakLegend}
           error={errors.arsak && texts.missingArsak}
         >
-          {Object.entries(oppfyltVurderingArsakTexts).map(
+          {Object.entries(unntakVurderingArsakTexts).map(
             ([arsak, text], index) => (
               <Radio
                 key={index}
@@ -99,7 +98,9 @@ export const OppfyltAktivitetskravSkjema = ({
         <BegrunnelseTextarea
           {...register("begrunnelse", {
             maxLength: begrunnelseMaxLength,
+            required: true,
           })}
+          error={errors.begrunnelse && texts.missingBegrunnelse}
           value={watch("begrunnelse")}
           label={texts.begrunnelseLabel}
         />
