+++ test/aktivitetskrav/AktivitetskravHistorikkTest.tsx
@@ -4,7 +4,7 @@ import { render, screen, within } from "@testing-library/react";
 import { ValgtEnhetContext } from "@/context/ValgtEnhetContext";
 import { navEnhet } from "../dialogmote/testData";
 import React from "react";
-import { AktivitetskravHistorikk } from "@/components/aktivitetskrav/historikk/AktivitetskravHistorikk";
+import { AktivitetskravHistorikk } from "@/sider/aktivitetskrav/historikk/AktivitetskravHistorikk";
 import {
   AktivitetskravStatus,
   AktivitetskravVarselDTO,
